// site/@cliotech/fx-core-api/src/facade/enums/IWebDialogType.ts
export enum IWebDialogType {
  NewDealSlip = "NewDealSlipDiv",
  NewLimit = "dlgNewLimit",
  TransactionSwitcher = "TransactionSwitcherDiv",
  NewPriceAlert = "NewPriceAlertDiv",
  EditLimitExpirationDate = "dlgEditLimitExpirationDate",
  NewDealValueDate = "dlgNewDealExpirationDate",
  EditClosingLimit = "dlgEditClosingLimit",
  CloseDeal = "dlgCloseDeal",
  ClosedDeal = "dlgClosedDeals",
  RolledOver = "RolledOverDiv",
  EditLimit = "dlgEditLimit",
  MarginStatus = "MarginStatusDiv",
  AmlStatus = "AmlStatusDiv",
  RelatedPosition = "RelatedPositionDiv",
  TradingSignalsTutorials = "TradingSignalsTutorialsDiv",
  EducationalTutorialsPopup = "EducationalTutorialsPopupDiv",
  MissingCustomerInformation = "MissingCustomerInformationDiv",
  MissingCustomerInformationContainer = "#paymentTypesContainerABTesting",
  NetExposuresSummaryDialog = "fx-component-netexposures-summary",
  CashBackDialog = "CashBackDiv",
  AccountCardRecords = "AccountCardRecordsDiv",
  ContractRollover = "ContractRolloverDiv",
  ClientQuestionnaire = "KycStatusDiv",
  DepositConfirmation = "DepositConfirmationDiv",
}