// site/@cliotech/fx-core-api/src/facade/enums/IWebViewType.ts
export enum IWebViewType {
  vQuotes = 1,
  vOpenDeals = 2,
  vLimits = 3,
  vClosedDeals = 4,
  vWallet = 5,
  vNewDeal = 6,
  vClosedDealDetails = 7,
  vNewLimit = 9,
  vBalance = 10,
  vNetExposure = 11,
  vWithdrawal = 12,
  vQuotesPreset = 14,
  vDeposit = 15,
  vEditClosingLimit = 16,
  vContact = 17,
  vPendingWithdrawal = 18,
  vEditLimit = 20,
  vMenu = 21,
  vCreditCardRegular = 22,
  vCharts = 23,
  vAddInstrument = 24,
  vTradingSummery = 25,
  vSummeryView = 26,
  vHomepageContent = 27,
  vSideMenu = 28,
  vChangePassword = 29,
  vPaymentTypes = 30,
  vActivityLog = 31,
  vAccountStatment = 33,
  vPersonalDetails = 34,
  vToolBar = 35,
  vPageTitle = 36,

  vClientQuestionnaire = 38,
  vClosedDealsSummaries = 39,
  vCloseDeal = 40,
  vRolledOver = 41,
  vOrderTabs = 43,
  vMarginStatus = 44,
  vAmlStatus = 45,
  vViewAndPrintWithdrawal = 46,
  vTransactionsReport = 47,
  vRelatedPosition = 48,
  vUploadDocuments = 49,
  vComodoSSLImage = 50,
  vTradingSignals = 52,
  vThirdParty = 53,
  vTradingSignalsTutorials = 54,
  vTutorials = 59,
  vCreditCard3rdParty = 60,
  vEditLimitExpirationDate = 61,
  vEmpCashU = 62,
  vEducationalTutorials = 63,
  vEducationalTutorialsPopup = 64,
  vEpaylinks = 65,
  vMoneyBookers = 72,
  vConcretePaymentForm = 75,
  vSafeChargeCashier = 100,
  vAdvinionChart = 101,
  vMissingCustomerInformation = 102,
  vAstropay = 106,
  vCddFinancialDetails = 107,
  vNetellerGO = 108,
  vPaySafeCard = 109,
  vWireTransfer = 110,
  vAccountSummaryWallet = 111,
  vNewDealSlip = 1000,
  vEconomicCalendar = 112,
  vMarketClosed = 115,
  vCashBack = 116,
  vGlobalCollectAPM = 117,
  vInatecAPMGiropay = 118,
  vInatecAPMPrzelewy24 = 119,
  vAccountCardRecords = 120,
  vNoAPIClearer = 121,
  vEcoPayz = 122,
  vSofortDirect = 123,
  vTransactionSwitcher = 125,
  vSignalsTechnicalAnalysisGrid = 126,
  vSignalsAlertGrid = 127,
  vSignalsCandleStickGrid = 128,
  vSignalsService = 129,
  vSignalsDisclamer = 130,
  vKluwp = 131,
  vContractRollover = 132,
  vDealsTabs = 133,
  vNotificationsSettings = 134,
  vClosedDealsDialog = 135,
  vRPN = 136,
  vPayPal = 137,
  vAstropayCard = 138,
  vGenericCreditCard = 139,
  vPayRetailers = 140,
  vDepositSuccessThankYou = 219,
  vDepositPendingThankYou = 220,
  vDepositConfirmation = 221,
  vSettings = 222,
  vAccountPreferences = 223,
  vEnableTrading = 224,
  vNewDealValueDate = 225,
  vPriceAlerts = 226,
  vRisingSunWeChat = 227,
  vNewPriceAlert = 228,
  vWireTransferSuccess = 229,
  vJeton = 230,
  vZotapayAPM = 231,
  vWithdrawalThankYou = 232,
}