// site/@cliotech/fx-core-api/src/facade/enums/IWebEForms.ts
export enum IWebEForms {
  Deals = 2,
  Statement = 5,
  Deposit = 6,
  Withdrawal = 7,
  ActivityLog = 8,
  PersonalDetails = 9,
  ClientQuestionnaire = 10,
  ChangePassword = 11,
  ViewAndPrintWithdrawal = 12,
  TransactionsReport = 13,
  UploadDocuments = 14,
  TradingSignals = 15,
  ThirdParty = 16,
  Tutorials = 19,
  EducationalTutorials = 20,
  AdvinionChart = 21,
  EconomicCalendar = 23,
  AccountCardRecords = 24,
  ContractRollover = 25,
  NotificationsSettings = 26,
  ConcretePaymentForm = 34,
  DepositSuccess = 64,
  DepositPending = 65,
  DepositConfirmation = 66,
  RolledOver = 67,
  Settings = 68,
  PriceAlerts = 69,
  WireTransferSuccess = 70,
  PendingWithdrawal = 71,
  WithdrawalThankYou = 72,
}